fix(wggo-phase2): seed adjoint with upstream dy + Pipe fall-through + u32 offset
Issue 1 (CRITICAL): emit_model_backward_bridge accepted dy_handle but bound it
to _dy_handle and discarded it. The adjoint loss seed defaulted to Constant(1.0),
corrupting per-head gradient magnitudes and potentially head ordering used by WGGO
pruning. Fix: expose AdjointGenerator::loss_seed_var_id(), add a pre-mapped-constant
check to PrimalOp::Constant arm in wengert_lower.rs, and pre-seed full_vars with
dy_handle before the adjoint compile_wengert_ops call (spec §4.2).

Issue 2: Pipe arm in WengertExtractor::extract_expr silently dropped free-function
RHS when resolved was None (e.g. x |> relu |> q_proj), collapsing to stub. Fix:
fall through to extract_expr(right) when the model-field path returns None.

Issue 3: emit_splice_memcpy offset parameter widened to u64 unnecessarily; the
GradArenaLayout contract uses u32. Narrowed to u32 to remove the footgun; all
call sites already passed u32 values (cast removed from binary_codegen.rs and
expr/mod.rs).

Issue 4: _arena_layout parameter in emit_model_backward_bridge now documents its
reserved purpose for Task 17+ cross-arena consistency checks.

All 181 calibration lib tests, 38 source_ad lib tests, and both integration test
suites (wggo_grad_arena_splice, wggo_refusals) green. Clippy lib clean.

+++ crates/nsl-codegen/src/calibration/retention.rs
@@ -25,7 +25,7 @@ use crate::calibration::observation::ProjectionRef;
 pub fn emit_splice_memcpy(
     fb: &mut FunctionBuilder,
     arena_ptr: Value,
-    offset: u64,
+    offset: u32,
     src_ptr: Value,
     nbytes: u64,
 ) {
